Updated dependency directory structure expected
The build structure expected by Libevent, qpid and gtest were all
based on in-source builds on windows and often involved hacking
third party makefiles and code just to get them to build on
modern compilers in both 32 and 64 bit. These changes update the
directory structure expected to reflect standard "install"
structures used by latest versions of the third party dependencies
used by OpenMAMA.

It also introduces the use of property sheets for Visual Studio
builds to centralize the setting up of include and library
settings in one location.

Both scons and Visual Studio project files have been updated.

+++ mama/c_cpp/src/c/bridge/qpid/io.c
@@ -69,7 +69,7 @@ void*
 qpidBridgeMamaIoImpl_dispatchThread (void* closure);
 
 void
-qpidBridgeMamaIoImpl_libeventIoCallback (int fd, short type, void* closure);
+qpidBridgeMamaIoImpl_libeventIoCallback (intptr_t fd, short type, void* closure);
 
 
 /*=========================================================================
@@ -262,7 +262,7 @@ qpidBridgeMamaIoImpl_dispatchThread (void* closure)
 }
 
 void
-qpidBridgeMamaIoImpl_libeventIoCallback (int fd, short type, void* closure)
+qpidBridgeMamaIoImpl_libeventIoCallback (intptr_t  fd, short type, void* closure)
 {
     qpidIoEventImpl* impl = (qpidIoEventImpl*) closure;
